pyproject.toml
[project]
name = "moodeng"
version = "0.1.0"
description = "🦛 Monitor for moo deng the hippo in her youtube stream"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Ted Nyman", email = "t@ted.io" }
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.7.0",
    "pillow>=9.5.0",
    "yt-dlp>=2023.11.16",
    "twilio>=8.1.0",
    "pushbullet.py>=0.11.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.21.0",
    "requests>=2.28.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.5",
    "scipy>=1.10.0",
    "ultralytics>=8.0.0",
    "gitpython>=3.1.30",
]

[project.scripts]
moodeng = "moodeng.cli:main"

[project.gui-scripts]
moodeng-gui = "moodeng.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/moodeng"]